Revert "Revert "Yeah, so it's a printf()-debugging-littered mess, but like it says on the""
This reverts commit b5b7abff8cc6092d82f88c93f29b678fcb27dcfc.

Okay so properly merge in the changes from the OmgAllTheTestsPass branch, this
time all nice and tidied up, like I originally intended!  &:S

The algorithm now basically works, although it will tend to find some people
that are not optimally close based on the fact that the squiral algorithm
assumes that the distance between cells is equal when the latitude and longitude
offsets add up to the same number.